refactor: add common widget for BottomBar
This gets rid of a lot code duplication and makes it more
straight-forward to build new screens for new features.

Note that the Row in BottomBar now has MainAxisAlignment.spaceAround by
default, so it's not necessary anymore to wrap buttons in `Expanded`.
