+++ lib/src/view/game/archived_game_screen.dart
@@ -3,7 +3,6 @@ import 'package:dartchess/dartchess.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:lichess_mobile/src/constants.dart';
 import 'package:lichess_mobile/src/model/analysis/analysis_controller.dart';
 import 'package:lichess_mobile/src/model/game/archived_game.dart';
 import 'package:lichess_mobile/src/styles/styles.dart';
@@ -16,6 +15,7 @@ import 'package:lichess_mobile/src/view/game/game_result_dialog.dart';
 import 'package:lichess_mobile/src/view/settings/toggle_sound_button.dart';
 import 'package:lichess_mobile/src/widgets/adaptive_action_sheet.dart';
 import 'package:lichess_mobile/src/widgets/board_table.dart';
+import 'package:lichess_mobile/src/widgets/bottom_bar.dart';
 import 'package:lichess_mobile/src/widgets/bottom_bar_button.dart';
 import 'package:lichess_mobile/src/widgets/buttons.dart';
 import 'package:lichess_mobile/src/widgets/countdown_clock.dart';
@@ -240,109 +240,89 @@ class _BottomBar extends ConsumerWidget {
     final canGoBackward = ref.watch(canGoBackwardProvider(gameData.id));
     final gameCursor = ref.watch(gameCursorProvider(gameData.id));
 
-    return Container(
-      color: Theme.of(context).platform == TargetPlatform.iOS
-          ? null
-          : Theme.of(context).bottomAppBarTheme.color,
-      child: SafeArea(
-        top: false,
-        child: SizedBox(
-          height: kBottomBarHeight,
-          child: Row(
-            children: [
-              Expanded(
-                child: BottomBarButton(
-                  label: context.l10n.menu,
-                  onTap: () {
-                    _showGameMenu(context, ref);
-                  },
-                  icon: Icons.menu,
-                ),
-              ),
-              gameCursor.when(
-                data: (data) {
-                  return Expanded(
-                    child: BottomBarButton(
-                      label: context.l10n.mobileShowResult,
-                      icon: Icons.info_outline,
-                      onTap: () {
-                        showAdaptiveDialog<void>(
-                          context: context,
-                          builder: (context) =>
-                              ArchivedGameResultDialog(game: data.$1),
-                          barrierDismissible: true,
-                        );
-                      },
-                    ),
+    return BottomBar(
+      children: [
+        BottomBarButton(
+          label: context.l10n.menu,
+          onTap: () {
+            _showGameMenu(context, ref);
+          },
+          icon: Icons.menu,
+        ),
+        gameCursor.when(
+          data: (data) {
+            return Expanded(
+              child: BottomBarButton(
+                label: context.l10n.mobileShowResult,
+                icon: Icons.info_outline,
+                onTap: () {
+                  showAdaptiveDialog<void>(
+                    context: context,
+                    builder: (context) =>
+                        ArchivedGameResultDialog(game: data.$1),
+                    barrierDismissible: true,
                   );
                 },
-                loading: () => const SizedBox.shrink(),
-                error: (_, __) => const SizedBox.shrink(),
               ),
-              Expanded(
-                child: BottomBarButton(
-                  label: context.l10n.gameAnalysis,
-                  onTap: ref.read(gameCursorProvider(gameData.id)).hasValue
-                      ? () {
-                          final (game, cursor) = ref
-                              .read(
-                                gameCursorProvider(gameData.id),
-                              )
-                              .requireValue;
+            );
+          },
+          loading: () => const SizedBox.shrink(),
+          error: (_, __) => const SizedBox.shrink(),
+        ),
+        BottomBarButton(
+          label: context.l10n.gameAnalysis,
+          onTap: ref.read(gameCursorProvider(gameData.id)).hasValue
+              ? () {
+                  final (game, cursor) = ref
+                      .read(
+                        gameCursorProvider(gameData.id),
+                      )
+                      .requireValue;
 
-                          pushPlatformRoute(
-                            context,
-                            builder: (context) => AnalysisScreen(
-                              title: context.l10n.gameAnalysis,
-                              pgnOrId: game.makePgn(),
-                              options: AnalysisOptions(
-                                isLocalEvaluationAllowed: true,
-                                variant: gameData.variant,
-                                initialMoveCursor: cursor,
-                                orientation: orientation,
-                                id: gameData.id,
-                                opening: gameData.opening,
-                                serverAnalysis: game.serverAnalysis,
-                                division: game.meta.division,
-                              ),
-                            ),
-                          );
-                        }
-                      : null,
-                  icon: Icons.biotech,
-                ),
-              ),
-              Expanded(
-                child: RepeatButton(
-                  onLongPress:
-                      canGoBackward ? () => _cursorBackward(ref) : null,
-                  child: BottomBarButton(
-                    key: const ValueKey('cursor-back'),
-                    // TODO add translation
-                    label: 'Backward',
-                    showTooltip: false,
-                    onTap: canGoBackward ? () => _cursorBackward(ref) : null,
-                    icon: CupertinoIcons.chevron_back,
-                  ),
-                ),
-              ),
-              Expanded(
-                child: RepeatButton(
-                  onLongPress: canGoForward ? () => _cursorForward(ref) : null,
-                  child: BottomBarButton(
-                    key: const ValueKey('cursor-forward'),
-                    // TODO add translation
-                    label: 'Forward',
-                    showTooltip: false,
-                    onTap: canGoForward ? () => _cursorForward(ref) : null,
-                    icon: CupertinoIcons.chevron_forward,
-                  ),
-                ),
-              ),
-            ],
+                  pushPlatformRoute(
+                    context,
+                    builder: (context) => AnalysisScreen(
+                      title: context.l10n.gameAnalysis,
+                      pgnOrId: game.makePgn(),
+                      options: AnalysisOptions(
+                        isLocalEvaluationAllowed: true,
+                        variant: gameData.variant,
+                        initialMoveCursor: cursor,
+                        orientation: orientation,
+                        id: gameData.id,
+                        opening: gameData.opening,
+                        serverAnalysis: game.serverAnalysis,
+                        division: game.meta.division,
+                      ),
+                    ),
+                  );
+                }
+              : null,
+          icon: Icons.biotech,
+        ),
+        RepeatButton(
+          onLongPress: canGoBackward ? () => _cursorBackward(ref) : null,
+          child: BottomBarButton(
+            key: const ValueKey('cursor-back'),
+            // TODO add translation
+            label: 'Backward',
+            showTooltip: false,
+            onTap: canGoBackward ? () => _cursorBackward(ref) : null,
+            icon: CupertinoIcons.chevron_back,
           ),
         ),
-      ),
+        RepeatButton(
+          onLongPress: canGoForward ? () => _cursorForward(ref) : null,
+          child: BottomBarButton(
+            key: const ValueKey('cursor-forward'),
+            // TODO add translation
+            label: 'Forward',
+            showTooltip: false,
+            onTap: canGoForward ? () => _cursorForward(ref) : null,
+            icon: CupertinoIcons.chevron_forward,
+          ),
+        ),
+      ],
     );
   }
