+++ lib/src/view/watch/tv_screen.dart
@@ -14,6 +14,7 @@ import 'package:lichess_mobile/src/utils/l10n_context.dart';
 import 'package:lichess_mobile/src/view/game/game_player.dart';
 import 'package:lichess_mobile/src/view/settings/toggle_sound_button.dart';
 import 'package:lichess_mobile/src/widgets/board_table.dart';
+import 'package:lichess_mobile/src/widgets/bottom_bar.dart';
 import 'package:lichess_mobile/src/widgets/bottom_bar_button.dart';
 import 'package:lichess_mobile/src/widgets/buttons.dart';
 import 'package:lichess_mobile/src/widgets/countdown_clock.dart';
@@ -209,72 +210,54 @@ class _BottomBar extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    return Container(
-      color: Theme.of(context).platform == TargetPlatform.iOS
-          ? null
-          : Theme.of(context).bottomAppBarTheme.color,
-      child: SafeArea(
-        top: false,
-        child: SizedBox(
-          height: kBottomBarHeight,
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceAround,
-            children: [
-              Expanded(
-                child: BottomBarButton(
-                  label: context.l10n.flipBoard,
-                  onTap: () => _flipBoard(ref),
-                  icon: CupertinoIcons.arrow_2_squarepath,
-                ),
-              ),
-              Expanded(
-                child: RepeatButton(
-                  onLongPress: ref
-                          .read(tvControllerProvider(tvChannel, game).notifier)
-                          .canGoBack()
-                      ? () => _moveBackward(ref)
-                      : null,
-                  child: BottomBarButton(
-                    key: const ValueKey('goto-previous'),
-                    onTap: ref
-                            .read(
-                              tvControllerProvider(tvChannel, game).notifier,
-                            )
-                            .canGoBack()
-                        ? () => _moveBackward(ref)
-                        : null,
-                    label: 'Previous',
-                    icon: CupertinoIcons.chevron_back,
-                    showTooltip: false,
-                  ),
-                ),
-              ),
-              Expanded(
-                child: RepeatButton(
-                  onLongPress: ref
-                          .read(tvControllerProvider(tvChannel, game).notifier)
-                          .canGoForward()
-                      ? () => _moveForward(ref)
-                      : null,
-                  child: BottomBarButton(
-                    key: const ValueKey('goto-next'),
-                    icon: CupertinoIcons.chevron_forward,
-                    label: context.l10n.next,
-                    onTap: ref
-                            .read(
-                              tvControllerProvider(tvChannel, game).notifier,
-                            )
-                            .canGoForward()
-                        ? () => _moveForward(ref)
-                        : null,
-                    showTooltip: false,
-                  ),
-                ),
-              ),
-            ],
+    return BottomBar(
+      children: [
+        BottomBarButton(
+          label: context.l10n.flipBoard,
+          onTap: () => _flipBoard(ref),
+          icon: CupertinoIcons.arrow_2_squarepath,
+        ),
+        RepeatButton(
+          onLongPress: ref
+                  .read(tvControllerProvider(tvChannel, game).notifier)
+                  .canGoBack()
+              ? () => _moveBackward(ref)
+              : null,
+          child: BottomBarButton(
+            key: const ValueKey('goto-previous'),
+            onTap: ref
+                    .read(
+                      tvControllerProvider(tvChannel, game).notifier,
+                    )
+                    .canGoBack()
+                ? () => _moveBackward(ref)
+                : null,
+            label: 'Previous',
+            icon: CupertinoIcons.chevron_back,
+            showTooltip: false,
           ),
         ),
-      ),
+        RepeatButton(
+          onLongPress: ref
+                  .read(tvControllerProvider(tvChannel, game).notifier)
+                  .canGoForward()
+              ? () => _moveForward(ref)
+              : null,
+          child: BottomBarButton(
+            key: const ValueKey('goto-next'),
+            icon: CupertinoIcons.chevron_forward,
+            label: context.l10n.next,
+            onTap: ref
+                    .read(
+                      tvControllerProvider(tvChannel, game).notifier,
+                    )
+                    .canGoForward()
+                ? () => _moveForward(ref)
+                : null,
+            showTooltip: false,
+          ),
+        ),
+      ],
     );
   }
